refactor: 사용하지 않는 메서드, DTO 삭제 #110
