+++ src/main/java/Seoul_Milk/sm_server/domain/taxInvoiceValidation/dto/request/NonVerifiedTaxValidationRequestDTO.java
@@ -1,6 +1,5 @@
 package Seoul_Milk.sm_server.domain.taxInvoiceValidation.dto.request;
 
-import Seoul_Milk.sm_server.domain.taxInvoiceValidation.dto.TaxInvoiceInfo;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 import lombok.Getter;
