+++ src/main/java/Seoul_Milk/sm_server/domain/taxInvoiceValidation/service/TaxInvoiceValidationService.java
@@ -1,12 +1,11 @@
 package Seoul_Milk.sm_server.domain.taxInvoiceValidation.service;
 
-import Seoul_Milk.sm_server.domain.taxInvoiceValidation.dto.TestDTO;
 import Seoul_Milk.sm_server.domain.taxInvoiceValidation.dto.request.NonVerifiedTaxValidationRequestDTO;
 import Seoul_Milk.sm_server.domain.taxInvoiceValidation.dto.response.NonVerifiedTaxValidationResponseDTO;
 import Seoul_Milk.sm_server.login.entity.MemberEntity;
 import com.fasterxml.jackson.core.JsonProcessingException;
+
 import java.io.UnsupportedEncodingException;
-import java.util.List;
 import java.util.concurrent.ExecutionException;
 
 public interface TaxInvoiceValidationService {
@@ -15,5 +14,4 @@ public interface TaxInvoiceValidationService {
 
     String verifiedTaxValidation(String key)
             throws UnsupportedEncodingException, JsonProcessingException, InterruptedException;
-    List<TestDTO> create();
 }
